requirements.txt
# Streamlit アプリケーションのための依存関係

# メインフレームワーク
streamlit>=1.28.0

# データ処理・可視化
pandas>=1.5.0
plotly>=5.15.0

# データ形式・設定ファイル
PyYAML>=6.0

# API リクエスト
requests>=2.28.0

# 日時処理 (Python標準ライブラリのdatetimeを使用)
# 追加の日時ライブラリは不要

# ファイル・OS操作 (Python標準ライブラリのos, shutilを使用)
# 追加のファイル操作ライブラリは不要

# XML解析 (Python標準ライブラリのxml.etree.ElementTreeを使用)
# 追加のXML解析ライブラリは不要

# タイプヒント (Python 3.9+では標準ライブラリのtypingを使用)
# typing_extensions  # Python < 3.9の場合のみ必要